# Conflicts:
#	the-game/out/production/the-game/game/GameField.class
#	the-game/out/production/the-game/game/Music.class
#	the-game/out/production/the-game/game/drawers/GameStage$3.class

+++ the-game/src/game/characters/Tower.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 public abstract class Tower implements GameEntity
 {
-    private static double attackCooldown;              // Delayed time for each attack
+    private static int attackCooldown;              // Delayed time for each attack
     private int attackDamage;                       // Amount of health to reduce from enemies per attack
     // Maximum range the tower can attack
     private int towerLevel;                         // The higher level the tower is the more effective it is
@@ -40,11 +40,11 @@ public abstract class Tower implements GameEntity
 
     }
 
-    public static double getAttackCooldown() {
+    public static int getAttackCooldown() {
         return attackCooldown;
     }
 
-    public void setAttackCooldown(double attackCooldown) {
+    public void setAttackCooldown(int attackCooldown) {
         Tower.attackCooldown = attackCooldown;
     }
 
@@ -82,8 +82,6 @@ public abstract class Tower implements GameEntity
 
         bullet.add(bullets);
     }
-
-
 
     public int getTowerLevel()
     {
@@ -167,7 +165,6 @@ public abstract class Tower implements GameEntity
     public Hill getTowerHill() {
         return towerHill;
     }
-
 
     public Point2D getPos() {
         return pos;
